% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_alpha.R
\name{calc_alpha}
\alias{calc_alpha}
\title{Calculate Cronbach's alpha}
\usage{
calc_alpha(scl, use)
}
\arguments{
\item{scl}{a scale within a multiple scaled data frame.}

\item{use}{an optional string indicating how to deal with missing values, See
\code{use} in \code{\link[stats]{cor}} for details.}
}
\value{
A numeric vector of length one. It is used to assess the internal
  consistency of a scale.
}
\description{
\code{calc_rbar()} is an internal function. It calculates the
  average correlation between items for a gradually emerging scale.
}
