% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rbar.R
\name{calc_rbar}
\alias{calc_rbar}
\title{Calculate the average correlation}
\usage{
calc_rbar(scl, use)
}
\arguments{
\item{scl}{a scale of a multiple scaled data frame}

\item{use}{an optional string indicating how to deal with missing values, See
\code{use} in \code{\link[stats]{cor}} for details.}
}
\value{
A numeric vector of length one that reports the inter-item
  correlation.
}
\description{
\code{calc_rbar()} is an internal functions. It calculates the
  average correlation of a fragment or scale.
}
