% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rit.R
\name{calc_rit}
\alias{calc_rit}
\title{Calculate the corrected item-total correlation}
\usage{
calc_rit(scl, use)
}
\arguments{
\item{scl}{a scale from a multiple scaled data frame.}

\item{use}{an optional string indicating how to deal with missing values, See
\code{use} in \code{\link[stats]{cor}} for details.}
}
\value{
A numeric vector of length one. It is used to examine the coherence
  between item and overall score.
}
\description{
\code{calc_rit()} calculates the corrected item-total
  correlation of the item with the sum score (of all the other variables) of a
  scale or fragment.
}
\details{
Note, that this procedure uses a part-whole correction. Thus, the
  item is excluded from the calculation process.
}
