% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_comp.R
\name{check_comp}
\alias{check_comp}
\title{Test if any correlation is greater than a specified value}
\usage{
check_comp(x, mrit_min, use)
}
\arguments{
\item{x}{some arbitrary input to be checked.}

\item{mrit_min}{a numeric constant of length 1 to specify the marginal
corrected item-total correlation.}

\item{use}{an optional string to specify how missing values enter the
analysis. See \code{use} in \code{\link[stats]{cor}} for details.}
}
\value{
There is no return value. The function is called for its side
  effects.
}
\description{
\code{check_comp()} is an internal function. \code{check_comp()}
  examines the correlation matrix, \code{cor(df)}. It complains (throws an
  error) if no correlation in \code{cor(df)} is greater than the specified
  \code{mrit_min}.
}
