% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_df.R
\name{check_df}
\alias{check_df}
\title{Test the appropriateness of an input}
\usage{
check_df(x)
}
\arguments{
\item{x}{some arbitrary input to be checked.}
}
\value{
There is no return value. The function is called for its side
  effects.
}
\description{
\code{check_df()} is an internal function which is called for
  its side effects. It guarantees that \code{x} is an appropriate data frame
  for the analysis. That mean: It verifies that \code{x} has less than two
  variables (a single item can't build a core), \code{x} has column names
  (used to pre-build \code{scls} in the overlapping process), if the column
  names are unique, and not of type \code{NA}. It throws an error if any of
  these requirements are not met. Additionally, it warns the user if
  the provided \code{colnames} are not unique or \code{NA}.
}
