% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mrit.R
\name{check_mrit}
\alias{check_mrit}
\title{Test the appropriateness of an input}
\usage{
check_mrit(x)
}
\arguments{
\item{x}{some arbitrary input to be checked.}
}
\value{
There is no return value. The function is called for its side
  effects.
}
\description{
\code{check_mrit()} is an internal function which is called for
  its side effects. It guarantees that the input is a double vector of length
  one. Moreover, the function secures that the lower bound is unique and
  ranges between `0` and `1` (it throws an error if not). In addition, it
  warns the user when pre-determining fragments.
}
