% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_msdf.R
\name{check_msdf}
\alias{check_msdf}
\title{Test the appropriateness of an input}
\usage{
check_msdf(x)
}
\arguments{
\item{x}{some arbitrary input to be checked.}
}
\value{
There is no return value. The function is called for its side
  effects.
}
\description{
\code{check_msdf()} is an internal function which is called for
  its side effects. It throws an error if the input is not of type `msdf`.
}
