% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_neg.R
\name{check_neg}
\alias{check_neg}
\title{Test the appropriateness of an input}
\usage{
check_neg(x)
}
\arguments{
\item{x}{some arbitrary input to be checked.}
}
\value{
There is no return value. The function is called for its side
  effects.
}
\description{
\code{check_neg()} is an internal function which is called for
  its side effects. It throws an error if the (\code{negative_too} input is
not a logical constant of length 1 and/or a missing value.
}
\details{
To verify both, that the (\code{negative_too} input is a logical
  constant of length 1 and not a missing value, is necessary because objects
  of type \code{NA} are logical constants of length 1, too.
}
