% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ovlp.R
\name{check_ovlp}
\alias{check_ovlp}
\title{Test the appropriateness of an input}
\usage{
check_ovlp(x)
}
\arguments{
\item{x}{some arbitrary input to be checked.}
}
\value{
There is no return value. The function is called for its side
  effects.
}
\description{
\code{check_ovlp()} is an internal function which is called for
  its side effects. It safeguards that\code{x}, is a character vector of
  length `1`. That means, it throws an error if not.
}
\details{
Within \code{disjoint()} and \code{overlap()}) \code{switch()}
  validates the input string itself, throwing an error message when the given
  character doesn't match any available option.
}
