% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sclvals.R
\name{check_sclvals}
\alias{check_sclvals}
\title{Test the appropriateness of an input}
\usage{
check_sclvals(x)
}
\arguments{
\item{x}{some arbitrary input to be checked.}
}
\value{
There is no return value. The function is called for its side
  effects.
}
\description{
\code{check_sclvals()} is an internal function which is called
  for its side effects. It checks whether \code{x} is a two element vector
  and throws an error if not. Integers are coerced to be of type
  \code{double}. Additionally, the function makes sure that the first value
  is not greater than the second.
}
\details{
Note that, when it checks whether \code{x} is a two element vector
  it implicitly assures that it is not \code{NULL}, since \code{NULL} is
  a logical constant of length `0`.
}
