% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_sclvals.R
\name{compare_sclvals}
\alias{compare_sclvals}
\title{Test the appropriateness of the relation between two inputs}
\usage{
compare_sclvals(x, x_attr)
}
\arguments{
\item{x}{some arbitrary input to be checked.}

\item{x_attr}{a numeric vector of length 2 indicating the start- and
endpoint.}
}
\value{
There is no return value. The function is called for its side
  effects.
}
\description{
\code{compare_sclvals()} is an internal function which is called
  for its side effects. It ensures that the \code{sclvals} set with
  \code{overlap()} are equal to those set with \code{disjoint()} -- and
  throws an error if not.
}
