% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disj_pci.R
\name{disj_pci}
\alias{disj_pci}
\title{Multiple Scaling In A Disjoint Manner Using Positive Items}
\usage{
disj_pci(df, mrit_mit, use)
}
\arguments{
\item{df}{a data frame object.}

\item{use}{an optional string to specify how missing values will enter the
analysis. See \code{use} in \code{\link[stats]{cor}} for details.}

\item{mrit_min}{a numerical constant to specify the marginal corrected item
total correlation. The value must be in the range of 0-1.}
}
\description{
\code{disj_pci} is an internal function and returns a list of
 data frames. In addition, \code{disj_pci} sets up an empty list of scales
 (\code{scls}) and scaling begins: (1) A (first) scale is set up. (2)
 \code{disj_pci} tries to find the two items with the highest correlation
 in the data frame. (3) If the this correlation is greater than the
 pre-specified lower bound (\code{mrit_min}), \code{disj_pci} adds up the two
 items to build the core of the scale. After developing the core, both items
 are excluded from the data frame. They are now part of the first scale. (4)
 While the correlation between the sum score and a remaining items of the
 data frame is greater than the pre-specified lower bound, \code{disj_pci}
 enlarges the scale with the appropriate item. (5 = 1) If there are at least
 two items left in the data frame, yielding a correlation coefficient greater
 than the set lower bound, \code{disj_pci} starts over again.
}
\details{
\code{disj_pci} build scales from items which share a positive
  linear relationship. Therefore, it tags the building block for the upcoming
  extensions -- including disjoint and overlapping scaling processes.

  \code{use} clarifies how to set up a correlation matrix in the presence of
  missing values. In a typical scaling process this happens at least twice.
  First, when determining the core items (the two items in the correlation
  matrix with the highest linear relationship). Second, when an item is
  proposed for an emerging scale.
}
\references{
Müller-Schneider, Thomas. (2001). Multiple Skalierung nach dem
  Kristallisationsprinzip / Multiple Scaling According to the Principle of
  Crystallization. Zeitschrift für Soziologie. 30. 10.1515/zfsoz-2001-0404.
}
