% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disjoint.r
\name{disjoint}
\alias{disjoint}
\title{Multiple scaling in a disjoint manner}
\usage{
disjoint(
  df,
  rit_min = 0.3,
  negative_too = FALSE,
  sclvals = NULL,
  use = "pairwise.complete.obs"
)
}
\arguments{
\item{df}{a data frame.}

\item{rit_min}{a corrected item total correlation between 0 and 1 marking the
lower bound of the scaling process.}

\item{negative_too}{a logical constant indicating whether negatively
correlating items should be included in the scaling process.}

\item{sclvals}{a double vector of the form \code{c(min,max)} marking the
start and end point of your scale, e.g.}
}
\description{
Multiple scaling in a disjoint manner
}
