% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_core.R
\name{extr_core}
\alias{extr_core}
\title{Extract the first two elements of an object}
\usage{
extr_core(scl)
}
\arguments{
\item{scl}{a scale from a multiple scaled data frame}
}
\value{
A numeric vector of length two. It contains the two items with the
  highest correlation in a fragment or scale.
}
\description{
\code{extr_core()} is used to extract all pairs of core items
  from a fragment.
}
