% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_core_nms.R
\name{extr_core_nms}
\alias{extr_core_nms}
\title{Extract the names of two core items}
\usage{
extr_core_nms(scl)
}
\arguments{
\item{scl}{a scale within a multiple scaled data frame}
}
\value{
A character vector of length two. It contains the names of the
  two items with the highest correlation in a fragment or scale.
}
\description{
\code{extr_core_nms()} is an internal function. It is used to
  extract the names of the core items.
}
