% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreb_itms.R
\name{extreb_itms}
\alias{extreb_itms}
\title{Extract everything but the core}
\usage{
extreb_itms(df, itm_nms)
}
\arguments{
\item{df}{a data frame object.}

\item{itm_nms}{the names of an item from a scale.}
}
\value{
A numeric vector of length m minus two, where 'm' specifies the
  number of variables in a given data set. It contains the items of a
  scale's counterpart.
}
\description{
\code{extreb_itms()} builds the counterpart of a fragment from
  the given item names. Therefore, the counterpart includes all variables
  that are not part of a fragment but which are mentioned in the specified
  data set.
}
