% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.msdf.R
\name{is.msdf}
\alias{is.msdf}
\title{Test for a multiple scaled data frame object}
\usage{
is.msdf(x)
}
\arguments{
\item{x}{an arbitrary object.}
}
\value{
a logical vector of length one. \code{TRUE} indicates that the object
  is of type \code{msdf}.
}
\description{
\code{is.msdf()} tests if the given object is a multiple scaled
  data frame.
}
\details{
\code{is.msdf} returns \code{TRUE} if the given object is of type
  \code{msdf} and FALSE otherwise.
}
