% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_muscldf.r
\name{new_muscldf}
\alias{new_muscldf}
\title{This so called \code{muscldf} is nothing more
  than a list of scales (in this case genuine subsets of that data frame).}
\usage{
new_muscldf(x = list(), method, rit_min, negative_too, sclvals = NULL, df)
}
\description{
This so called \code{muscldf} is nothing more
  than a list of scales (in this case genuine subsets of that data frame).
}
