% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nme_msdf.R
\name{nme_msdf}
\alias{nme_msdf}
\title{Rename the components of a multiple scaled data frame}
\usage{
nme_msdf(x)
}
\arguments{
\item{x}{a multiple scaled data frame.}
}
\value{
A character vector that numbers each element of its input according
  to the naming scheme described above.
}
\description{
\code{nme_msdf()} is an internal function. It renames the
  components of a msdf
}
\details{
The naming convention is \code{scl_n}. \code{scl} stands for
  `scale` and \code{n} is the number of the fragment or scale. For example,
  the first component is called \code{scl_1}.
}
