% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{Multiple Scaling In An Overlapping Manner}
\usage{
overlap(
  msdf,
  mrit_min = NULL,
  negative_too = FALSE,
  overlap_with = "fragment",
  sclvals = NULL,
  use = "pairwise.complete.obs"
)
}
\arguments{
\item{msdf}{a multiple scaled data frame (built with \code{disjoint}).}

\item{mrit_min}{a numerical constant of length 1 to specify the marginal
corrected item-total correlation. It's value is in the range of 0-1. The
default is set to \code{.3}.}

\item{negative_too}{a logical constant indicating whether reversed items
should be included. The default is set to \code{FALSE}.}

\item{overlap_with}{a string telling \code{overlap} the items it should
extend on (in each particular case). To build up on all variables of a
fragment use \code{fragment}, for the cores-only option type \code{core}.
The default is set to "fragment".}

\item{sclvals}{a numerical vector of length 2 indicating the start and
endpoint of a scale. Use something like \code{c(min,max)}.}

\item{use}{an optional string to specify how missing values will enter the
analysis. See \code{use} in \code{\link[stats]{cor}} for details. The
default is set to \code{pairwise.complete.obs}.}
}
\description{
\code{overlap} returns a multiple, overlapping scaled version of
  the specified \code{msdf}.
}
\details{
\code{use} clarifies how to set up a correlation matrix in the
  presence of missing values. In a typical scaling process this happens at
  least twice. First, when determining the core items (the two items in the
  correlation matrix with the highest linear relationship). Second, when an
  item is proposed for an emerging scale.
}
\examples{
# Build a msdf
msdf <- disjoint(mtcars, mrit_min = .4)

# Using positive correlations (and `pairwise.complete.obs`)
overlap(msdf, mrit_min = .6, overlap_with = "core")

# Including negative correlations (and `pairwise.complete.obs`)
overlap(msdf, mrit_min = .7, negative_too = TRUE, sclvals = c(-3,3))

# Changing the treatment of missing values
overlap(msdf, mrit_min = .6, use = "all.obs")

}
\references{
Müller-Schneider, Thomas. (2001). Multiple Skalierung nach dem
  Kristallisationsprinzip / Multiple Scaling According to the Principle of
  Crystallization. Zeitschrift für Soziologie. 30. 10.1515/zfsoz-2001-0404.
}
