% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovlp_nci.R
\name{ovlp_nci}
\alias{ovlp_nci}
\title{Multiple Scaling In An Overlapping Manner Using Negative Items Too}
\usage{
ovlp_nci(msdf, mrit_min, overlap_with, sclvals, use)
}
\arguments{
\item{msdf}{a multiple scaled data frame (built with \code{disjoint}).}

\item{mrit_min}{a numerical constant of length 1 to specify the marginal
corrected item-total correlation. It's value is in the range of 0-1.}

\item{overlap_with}{a string telling \code{overlap} the items it should
extend on (in each particular case). To build up on all variables of a
fragment use \code{fragment}, for the cores-only option type \code{core}.}

\item{sclvals}{a numerical vector of length 2 indicating the start and
endpoint of a scale. Use something like \code{c(min,max)}.}

\item{use}{an optional string to specify how missing values will enter the
analysis. See \code{use} in \code{\link[stats]{cor}} for details.}

\item{negative_too}{a logical constant indicating whether items with a
negative correlation should be included in the scaling process.}
}
\description{
\code{ovlp_nci} is an internal function and returns a list of
  data frames. It takes a disjoint (\code{msdf}) and tries to extend it with
  those items in the data frame which are not yet built into the fragment
  (aka., counterpart). Despite the default, one can extend on the cores of
  each fragment. The underlying expansion algorithm remains the same. The
  only difference is that \code{ovlp_nci} applies this algorithm to each
  fragment within `msdf`. The overlapping family of functions is thus a
  multiple extension of the disjoint scaling approach to various fragments.
  In addition, \code{ovlp_nci} is capable of soaking up reversed items.
}
\references{
Müller-Schneider, Thomas. (2001). Multiple Skalierung nach dem
  Kristallisationsprinzip / Multiple Scaling According to the Principle of
  Crystallization. Zeitschrift für Soziologie. 30. 10.1515/zfsoz-2001-0404.
  @importFrom stats cor
}
