% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.muscldf.r
\name{print.muscldf}
\alias{print.muscldf}
\title{Print method for multiple scaled data frame (msdf) objects}
\usage{
\method{print}{muscldf}(muscldf, digits = 2, use = "pairwise.complete.obs")
}
\arguments{
\item{muscldf}{a multiple scaled data frame (built with either
\code{disjoint} or \code{overlap}).}

\item{use}{an optional string indicating how to deal with missing values if
necessary. See \code{use} in \code{\link[stats]{cor}} for details. The
default is set to `pairwise.complete.obs`.}
}
\description{
The print mehtod for msdf objects returns a list of summary
  statistics. It is tailored for the use with the multiple scaling approach
  and returns the marginal corrected item total correlation (`mrit`),
  Cronbachs Alpha (`alpha`) and reports the average correlation (`rbar`) for
  every scale at each step of the scaling process.
}
\details{
The \code{use} argument takes control over the treatment of missing
  values when correlation matrices are built. In a (usual) scaling process
  this happens at least twice: first when determining the core (the two items
  of the correlation matrix with the highest linear relationship), and second
  when an item is considered to be part of a given scale. Since the print is
  programmed to reconstructs the scale building process (done with
  \code{disjoint} or \code{overlap}) to get its summary statistics, one
  should stay with the specified options.
}
\examples{
# Exploring a disjoint scaled data frame (`muscldf`)
muscldf <- disjoint(mtcars, rit_min = .4)
explore(muscldf)

# Exploring an overlapping scaled data frame (`msdf_ovlp`)
msdf_disj <- disjoint(mtcars, rit_min = .4)
msdf_ovlp <- overlap(msdf_disj, rit_min = .7, overlap_with = "full_scale")
explore(msdf_ovlp)
}
\references{
Müller-Schneider, Thomas. (2001). Multiple Skalierung nach dem
  Kristallisationsprinzip / Multiple Scaling According to the Principle of
  Crystallization. Zeitschrift für Soziologie. 30. 10.1515/zfsoz-2001-0404.
}
