% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvrs_note.R
\name{rvrs_note}
\alias{rvrs_note}
\title{Notification message for a reversed variable}
\usage{
rvrs_note(msg, applicant)
}
\arguments{
\item{msg}{a reverse message sent from either \code{disjoint()} or
\code{overlap()}}

\item{applicant}{the function which wants to leave messages \code{disjoint()}
or \code{overlap()})}
}
\value{
The function is called for its side effects. It leaves a message when
  an item is reversed.
}
\description{
\code{rvrs_note()} gets the full report of reversed variables
  and reports a unique list of them.
}
