% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvrs_var.R
\name{rvrs_var}
\alias{rvrs_var}
\title{Reverse a variable}
\usage{
rvrs_var(var, sclvals)
}
\arguments{
\item{var}{a variable or item (often a column from a data frame).}

\item{sclvals}{the start and end point of a scale (specify: \code{c(sp,ep)})}
}
\value{
A reversed numeric vector.
}
\description{
\code{rvrs_var()} reverses an item using the specified scaling
  values. It handles the following types of scales:
  \itemize{
  \item  ...-3 -2 -1 0 1 2 3..., e.g., \code{sclvals = c(-3, 3)}
  \item  0 1 2 3 4 5 6..., e.g.,  \code{sclvals = c(0, 7)}
  \item  1 2 3 4 5 6 7..., e.g., \code{sclcals = c(1, 7)}
  }
}
