% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trust.r
\docType{data}
\name{trust}
\alias{trust}
\title{German General Social Survey's Trust Items}
\format{
A subset of the German General Social Survey (ALLBUS) 2018 containing
  its `trust items` (F018). Participants were asked about the trust they
  place in public institutions and organizations. `1` means they have
  absolutely no trust at all, whereas `7` represents a great deal of trust.
  The data frame has 3477 rows and 13 variables: \describe{
  \item{healserv}{Health service}
  \item{fccourt}{German constitutional court}
  \item{bundtag}{German Parliament}
  \item{munadmin}{Municipal administration}
  \item{judsyst}{Judicial system}
  \item{tv}{Television}
  \item{newsppr}{Newspapers}
  \item{uni}{Universities and other institutes of higher education}
  \item{fedgovt}{}
  \item{police}{}
  \item{polpati}{}
  \item{eucomisn}{}
  \item{eupalmnt}{}
}
Note: Participant order has randomly rearranged to ensure their anonymity.
}
\usage{
trust
}
\description{
German General Social Survey's Trust Items
}
\references{
{ GESIS - Leibniz-Institut für Sozialwissenschaften (2019).
German General Social Survey - ALLBUS 2018. GESIS Datenarchiv, Köln. ZA5272
Datenfile Version 1.0.0, https://doi.org/10.4232/1.13325.}
}
\keyword{datasets}
