% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checks}
\alias{checks}
\alias{check_df}
\alias{check_sclvals}
\alias{compare_sclvals}
\alias{check_mrit}
\alias{check_ovlp}
\alias{check_msdf}
\alias{check_neg}
\alias{check_comp}
\title{A Set of Checks for the Appropriteness of Inputs}
\usage{
check_df(x)

check_sclvals(x)

compare_sclvals(x, x_attr)

check_mrit(x)

check_ovlp(x)

check_msdf(x)

check_neg(x)

check_comp(x, mrit_min, use)
}
\arguments{
\item{x}{some arbitrary input to be checked}
}
\description{
A Set of Checks for the Appropriteness of Inputs
}
\details{
\code{check_df(x)}, checks whether \code{x} is a data frame, if it has less
  than two variables (a single item can't set up a core), whether \code{x}
  has column names (used to pre-build \code{scls} in the overlapping
  process), if column names are unique and not \code{NA}.

  \code{check_sclvals(x)} checks whether \code{x} is a two element vector
  (implicitly assures that it is not \code{NULL}, since \code{NULL} is a
  logical constant of length `0`.). Integers are coerced to be of type
  \code{double}. Additionally the function makes sure that the first value is
  not greater than the second.

  \code{compare_sclvals(x)} is used when the user sets \code{sclvals} in an
  overlapping process, but there are already \code{sclvals} given from the
  disjoint scaling process done before. The function assures that the
  attribute is identical to the input (it doesn't make sense to change
  \code{sclvals} in the process of analysis.)

  \code{check_rit(x)} makes sure that the input is a double vector of length
  one (a user can only specify one lower bound in a scaling process) and in
  the range between `0` and `1`.

  \code{check_ovlp(x)} assures that the method to `overlap_with`, i.e.
  \code{x}, is a character vector of length `1`. Any further checks are
  required since \code{switch} (used in the main functions \code{disjoint}
  and \code{overlap}) validates the input strings itself, throwing an error
  message when the given character doesn't match the available options

  \code{check_msdf(x)} is used in \code{explore} and \code{overlap} to make
  sure the specified input is of type `msdf`. Hint: No further checks for
  msdf necessary. A data frame is only scaled when it has names see:
  \code{check_df}. Since (1) \code{disjoint} and \code{overlap} do not
  manipulate the `colnames` attribute and the user cant specify an
  alternative \code{df}`, names must  be valid.

  \code{check_neg(x)} checks the negative_too input is a logical constant of
  length 1. Because any \code{NA} value is a logical constant of length `0`
  one must additionally assure that the input is not a \code{NA}.

  \code{check_comp} checks the compatibility of the specified \code{mrit_min}
  value and any pair of correlations in the given data frame. If the value of
  \code{mrit_min} is smallr it throws an error.
}
