% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disj_pci.R
\name{disj_pci}
\alias{disj_pci}
\title{Multiple scaling in a disjoint manner using positive items}
\usage{
disj_pci(df, mrit_min, use)
}
\arguments{
\item{df}{a data frame object.}

\item{mrit_min}{a numerical constant to specify the marginal corrected item
total correlation. The value must be in the range of 0-1.}

\item{use}{an optional string to specify how missing values will enter the
analysis. See \code{use} in \code{\link[stats]{cor}} for details.}
}
\value{
A list of data frames. It results from applying the above algorithm.
}
\description{
\code{disj_pci} is an internal function and returns a list of
  data frames. It is a loop which runs through the following steps: (1) Set
  up a (first) scale. (2) find the two items with the highest positive
  correlation in the data set. (3) If the absolute value of this correlation
  is greater than the pre-specified lower bound (\code{mrit_min}), add up the
  two items to build the core of the emerging scale. (4) As long as the value
  of the correlation between the sum score and a remaining items in the data
  frame is greater than \code{mrit_min}, flavor the scale with the
  appropriate item. (5) If there are at least two leftovers in the data
  frame, that meet the inclusion criterion, start over again.
}
\details{
\code{disj_pci} build scales from items which share a positive
  linear relationship. Therefore, it tags the building block for the upcoming
  extensions -- including disjoint and overlapping scaling processes.

  \code{use} clarifies how to set up a correlation matrix in the presence of
  missing values. In a typical scaling process this happens at least twice.
  First, when determining the core items (the two items in the correlation
  matrix with the highest linear relationship). Second, when an item is
  proposed for an emerging scale.
}
\references{
Müller-Schneider, T. (2001). Multiple Skalierung nach dem
  Kristallisationsprinzip / Multiple Scaling According to the Principle of
  Crystallization. Zeitschrift für Soziologie, 30(4), 305-315.
  https://doi.org/10.1515/zfsoz-2001-0404
}
