% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disjoint.r
\name{disjoint}
\alias{disjoint}
\title{Multiple Scaling In A Disjoint Manner}
\usage{
disjoint(
  df,
  rit_min = 0.3,
  negative_too = FALSE,
  sclvals = NULL,
  use = "pairwise.complete.obs"
)
}
\arguments{
\item{df}{a data frame object.}

\item{rit_min}{a numerical constant to specify the (corrected item total)
correlation. The value of this lower bound must in the range of 0 to 1. The
default is set to \code{.3}.}

\item{negative_too}{a logical constant indicating whether items with a
negative correlation should be included in the scaling process. The default
is set to \code{FALSE}.}

\item{sclvals}{a numerical vector of length 2 indicating the first and the
full scale value. Consider using the shape \code{c(min,max)}.}

\item{use}{an optional string indicating how to deal with missing values if
necessary. See \code{use} in \code{\link[stats]{cor}} for details. The
default is set to \code{"pairwise.complete.obs"}.}
}
\description{
\code{disjoint} returns a multiple, disjoint scaled version of
  the specified data frame. This so called \code{muscldf} is nothing more
  than a list of scales (in this case genuine subsets of that data frame).
}
\details{
The \code{use} argument takes control over the treatment of missing
  values when correlation matrices are build. In a scaling process this
  happens at least twice: first when determining the core (the two items of
  the correlation matrix with the highest linear relationship), and second
  when an item is considered to be part of this scale.

  Hint: If there ought to be two items in the scaling process, having an
  equal correlation, e.g. with the sum score, always the first one is
  used.
}
\examples{
# Using positive correlations (and `pairwise.complete.obs`)
disjoint(mtcars, rit_min = .4)

# Including negative correlations (and `pairwise.complete.obs`)
disjoint(mtcars, rit_min = .4, negative_too = TRUE, sclvals = c(1,7))

# Changing the treatment of missing values
disjoint(mtcars, rit_min = .4, use = "all.obs")
}
\references{
Müller-Schneider, Thomas. (2001). Multiple Skalierung nach dem
  Kristallisationsprinzip / Multiple Scaling According to the Principle of
  Crystallization. Zeitschrift für Soziologie. 30. 10.1515/zfsoz-2001-0404.
}
