% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_msdf.R
\name{new_msdf}
\alias{new_msdf}
\title{Creator function to build a new \code{msdf}}
\usage{
new_msdf(x = list(), method, mrit_min, negative_too, sclvals = NULL, df)
}
\description{
A \code{msdf} is a multiple scaled data frame. An object of that
  type is nothing more than a list of data frames with some attributes
  attached to it. The data frame itself is a subset of the given data set --
  arranged according to the crystallization principle.
}
\details{
Multiple scaled data frames are only for internal use. A test
  function ensures that the argument to \code{overlap} is a \code{msdf}.
}
