% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_muscldf.R
\name{new_muscldf}
\alias{new_muscldf}
\title{Creator Function to Build A New \code{muscldf}}
\usage{
new_muscldf(x = list(), method, mrit_min, negative_too, sclvals = NULL, df)
}
\description{
A \code{muscldf} is a multiple scaled data frame. A multiple
scaled data frame is nothing more than a list of scales with attributes. A
scale is just a subset of the specified data frame (e.g., \code{disjoint}) or
an extension of the given \code{muscldf} (e.g., \code{overlap} -- were each
scale is built according to the crystallization principle.
}
\details{
The \code{muscldf} is only for internal use. A check function
assures that the argument to \code{overlap} is of that type. Another useful
feature is that the attributes provide information which settings were used
in the scaling process -- after scaling has been done. So one can be sure
that all parameters are set correctly.
}
