% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils}
\alias{utils}
\alias{nme_muscldf}
\alias{calc_rbar}
\alias{calc_alpha}
\alias{calc_rit}
\alias{extr_itms}
\alias{extreb_itms}
\alias{rvrs_var}
\title{A Set of Urtility Functions}
\usage{
nme_muscldf(x)

calc_rbar(scl, use)

calc_alpha(scl, use)

calc_rit(scl, use)

extr_itms(df, itm_nms)

extreb_itms(df, itm_nms)

rvrs_var(var, sclvals)
}
\arguments{
\item{x}{a muscldf}

\item{scl}{a scale from a muscldf}

\item{use}{an optional string indicating how to deal with missing values if
necessary. See \code{use} in \code{\link[stats]{cor}} for details.}

\item{df}{a data frame object,}

\item{itm_nms}{the names of an item from a scale.}
}
\description{
A Set of Urtility Functions
}
\details{
\code{nme_muscldf} names a muscldf. It is used for convenience and
replaces the common list element indicator `[[1]]`.

\code{calc_rbar} calculates the average correlation between the items at a
certain point of the scaling process.
}
