% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{nme_muscldf}
\alias{calc_rbar}
\alias{calc_alpha}
\alias{calc_mrit}
\alias{extr_core}
\alias{extr_core_nms}
\alias{extreb_itms}
\alias{rvrs_var}
\title{A Set of Utility Functions}
\usage{
nme_muscldf(x)

calc_rbar(scl, use)

calc_alpha(scl, use)

calc_mrit(scl, use)

extr_core(scl)

extr_core_nms(scl)

extreb_itms(df, itm_nms)

rvrs_var(var, sclvals)
}
\arguments{
\item{x}{a muscldf}

\item{scl}{a scale from a muscldf}

\item{use}{an optional string indicating how to deal with missing values if
necessary. See \code{use} in \code{\link[stats]{cor}} for details.}

\item{df}{a data frame object,}

\item{itm_nms}{the names of an item from a scale.}

\item{var}{a variable or item (often a column from a data frame)}

\item{sclvals}{the start and end point of a scale (consider \code{c(sp,ep)})}
}
\description{
A Set of Utility Functions
}
\details{
\code{nme_muscldf} names a muscldf. It is used for convenience and
  to replace the common list element indicator, e.g.,`[[1]]`, by assigning a
  name to the scale, e.g., `scl_1`.

  \code{calc_rbar} calculates the average correlation between the items at a
  certain point of the scaling process.

  \code{calc_alpha} calculates Cronbachs Alpha of a scale (i.e., the internal
  consistency of the items) at a certain point of the scaling process.

  \code{calc_mrit} calculates the Corrected Item-Total Correlation of an item
  with the sum score of all the other variables of the scale -- at a certain
  point of the scaling process. It uses a part-whole correction, i.e., the
  item itself is excluded in the calculation process.

  \code{extr_itms} subsets a data frame given a set of item names from a
  scale built in the disjoint scaling process (i.e., from a muscldf). It is
  used to construct one of the `working frames` in the overlapping process.
  Being more precise, \code{extr_itms} collects (instructed by the
  `overlap_with` argument) if only the cores or the full_scale should be used
  as a base to overlap.
  DELETE IF TESTED!

  \code{extr_core} extracts (multiple) pairs of core items from a given data
  frame (`scl`), which is often part of a list of data frames nested within a
  `msdf` object. `overlap()` draws on this helper function when setting the
  `overlap_with=core` option. Thus, it forms the base for the upcoming
  overlap procedure.

  \code{extr_core_nms} extracts the name of the core item from a given data
  frame (`scl`), which is often part of a list of data frames nested within a
  `msdf` object. `overlap()` draws on this helper function when setting the
  `overlap_with=core` option. It litterally sets up the counterpart for the
  upcoming overlap procedure.

  \code{rvrs_var} reverses the item given the specified scaling values. It
  can handle the following types of scales:
  \itemize{
  \item  ...-3 -2 -1 0 1 2 3..., e.g., \code{sclvals = c(-3, 3)}
  \item  0 1 2 3 4 5 6..., e.g.,  \code{sclvals = c(0, 7)}
  \item  1 2 3 4 5 6 7..., e.g., \code{sclcals = c(1, 7)}
  }
}
